# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handlers.createGameRequest:
    properties:
      deckCount:
        example: 2
        maximum: 8
        minimum: 1
        type: integer
      maxPlayers:
        example: 8
        type: integer
      minPlayers:
        example: 2
        type: integer
    required:
    - deckCount
    type: object
...
  handlers.startGameRequest:
    properties:
      initialDealCount:
        example: 2
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: 'Card-game engine: shoe management, dealing, shuffling, scoring.'
  title: DeckForge API
  version: "1.0"
paths:
...
  /games:
    get:
      description: Returns all non-finished games, optionally filtered by status.
      parameters:
      - description: WAITING or IN_PROGRESS
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: games array with playerCount and dealerUsername
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List games
      tags:
      - games
    post:
      consumes:
      - application/json
      description: Creates a new game. The caller becomes the dealer and auto-joins
        as player seat 0. Initial shoe cards are added automatically for the requested
        deck count.
      parameters:
      - description: game config
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.createGameRequest'
      produces:
      - application/json
      responses:
        "201":
          description: game + player
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a game
      tags:
      - games
  /games/{id}:
    delete:
      description: Permanently removes a game. Allowed for the dealer of that game
        or any admin.
      parameters:
      - description: game UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: no content
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a game
      tags:
      - games
    get:
      description: Returns a game's current state, shoe status, and dealer username.
      parameters:
      - description: game UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: game + totalCards + remainingCards + dealerUsername
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get game detail
      tags:
      - games
...
  /games/{id}/end:
    post:
      description: Transitions an IN_PROGRESS game to FINISHED and returns the final
        state.
      parameters:
      - description: game UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: updated game
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: End a game
      tags:
      - games
  /games/{id}/players:
    get:
      description: Returns all active players sorted by total hand value descending.
        Ties broken by seat order ascending. Scores are intentionally public during
        play — cards remain private.
      parameters:
      - description: game UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: leaderboard array [{playerId, userId, username, seatOrder,
            handValue, cardCount}]
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get the leaderboard
      tags:
      - players
    post:
      description: Adds the authenticated user as a player. If the game is IN_PROGRESS
        a catch-up deal is applied automatically.
      parameters:
      - description: game UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: player
          schema:
            additionalProperties: true
            type: object
        "400":
          description: already in a game, game full, etc.
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: game not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Join a game
      tags:
      - players
  /games/{id}/players/{pid}:
    delete:
      description: The player themselves, the dealer, or an admin may remove a player.
        All held cards are returned to the shoe.
      parameters:
      - description: game UUID
        in: path
        name: id
        required: true
        type: string
      - description: player UUID
        in: path
        name: pid
        required: true
        type: string
      responses:
        "204":
          description: removed
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove a player from a game
      tags:
      - players
...
  /games/{id}/start:
    post:
      consumes:
      - application/json
      description: Transitions the game from WAITING to IN_PROGRESS and optionally
        deals an initial hand to each player.
      parameters:
      - description: game UUID
        in: path
        name: id
        required: true
        type: string
      - description: optional initial deal count per player
        in: body
        name: body
        schema:
          $ref: '#/definitions/handlers.startGameRequest'
      produces:
      - application/json
      responses:
        "200":
          description: updated game
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Start a game
      tags:
      - games
securityDefinitions:
  BearerAuth:
    description: JWT obtained from POST /auth/register or POST /auth/login.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"